requirements.txt
# PhotoBooth – Python backend dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
opencv-python>=4.9.0
Pillow>=10.2.0
python-multipart>=0.0.9

# Legacy PySide6 desktop app (app/)
PySide6
watchdog
qrcode[pil]
